+++ components/technical-doc/index.js
