+++ pages/docs/communes.js
@@ -3,7 +3,7 @@ import FaMapSigns from 'react-icons/lib/fa/map-signs'
 import Head from '../../components/head'
 import Page from '../../layouts/main'
 
-import Communes from '../../components/communes/communes'
+import Communes from '../../components/communes'
 
 const title = 'Communes'
 const description = 'Interrogez les référentiels géographiques plus facilement.'
