+++ pages/docs/regions.js
@@ -3,7 +3,7 @@ import FaMapSigns from 'react-icons/lib/fa/map-signs'
 import Head from '../../components/head'
 import Page from '../../layouts/main'
 
-import Regions from '../../components/regions/regions'
+import Regions from '../../components/regions'
 
 const title = 'Régions'
 const description = 'Interrogez les référentiels géographiques plus facilement.'
